chore: rename test utils to common to make room for tests against utilty functions
