+++ test/e2e/index.ts
@@ -29,7 +29,7 @@ import {
   launch,
   makeCoordinatorZomeBundle,
   withConductor,
-} from "./util.js";
+} from "./common.js";
 
 const delay = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
